projects/uranus/scripts/requirements.txt
bibtexparser==1.2.0
PyYAML==6.0.0